// src/notifications/constants/notification-types.ts
export enum NotificationType {
  RECEIPT_ADDED = 'RECEIPT_ADDED',
  SPENDING_GAP = 'SPENDING_GAP',
  REPORT_CREATED = 'REPORT_CREATED',
  REPORT_UPDATED = 'REPORT_UPDATED',
  REPORT_RESOLVED = 'REPORT_RESOLVED',
  COMMENT_ADDED = 'COMMENT_ADDED',
  TURN_STARTED = 'TURN_STARTED',
  TURN_REMINDER = 'TURN_REMINDER',
  CLEANING_TASK_ASSIGNED = 'CLEANING_TASK_ASSIGNED',
  CLEANING_TASK_COMPLETED = 'CLEANING_TASK_COMPLETED',
  TENANT_DISABLED = 'TENANT_DISABLED',
  TENANT_REACTIVATED = 'TENANT_REACTIVATED',
  ANNOUNCEMENT = 'ANNOUNCEMENT',
  MEMBER_JOINED = 'MEMBER_JOINED',
}